requirement.txt
tensorflow
pillow
matplotlib
faiss-cpu